msm: mdss: Enable dynamic clock gating for MDSS AHB clock
On APQ8084, currently AHB clock is always on from device tree settings.
With this change, it will be enabled if BWC decode is enabled and source
dimensions are changed. Since the source dimensions are double buffered,
the MDP AHB clock will have to be forced on before writing to the source
dimension registers and then the force on turned off after the changes
have been applied (i.e. next vsync).

Change-Id: Ie3e73217bf184a128a4cf677688fdc4019ae44d5
Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>

Conflicts:
	drivers/video/msm/mdss/mdss_mdp_pipe.c