hotplug: avoid load jumping for hotplug cpus
Imagine the following load scenario:

 CPU      0   1   2   3
 Load(%)  5   0   90  95

This happens when the scheduler decides to move processes
from the parent cpus (0,1) to the newley onlined child cpus.
In this use case, the hotplug algorithm would decide to
offline cpu 2, because its the lowest load cpu for the child
cpus.
Somehow this is a very bad thing, because first the process
was migrated to another cpu, then in the next sampling it would
be migrated back to another cpu simply because we want to offline
a cpu, which wastes time.
So this patch aims to prevent this jumping-effect by comparing
the parent and child load and stop offline operations for a
sampling, hopefully giving the scheduler the possibility to
balance the load nicely.
While this might not strike in every load scenario, it doesnt
harm in terms of battery usage.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>